fix(scale): 修复刻度解释文本格式及函数类型声明
- 删除各职业兴趣类型描述中的多余换行符，修正文本为连贯段落
- 规范estimated_duration函数中变量类型，去除多余as转换
- 修改DIMENSIONS常量类型声明，去除多余的'static生命周期标注，提高代码简洁性

+++ src/scale/mod.rs
@@ -329,8 +329,8 @@ impl ScaleListItem<'_> {
 ///
 /// 每个问题耗时 20 秒为标准计算
 const fn estimated_duration(total_questions: u32) -> [u32; 2] {
-    let min = (total_questions * 10 / 60) as u32;
-    let max = (total_questions * 30 / 60) as u32;
+    let min = total_questions * 10 / 60;
+    let max = total_questions * 30 / 60;
     [min, max]
 }
